stratiform: replace TelParser with the streaming implementation
Deletes the old buffered TelParser.scala (1393 lines) and renames the
streaming parser file + object + class to TelParser. All call sites in
stratiform.Tel.scala, the tel"..." interpolator macros, the test suite,
and the benchmark module are updated to refer to the unified TelParser.

The benchmark module loses its now-redundant "Parse (streaming)" entries
(they invoked the same code as "Parse (no schema)" after the swap) and
the streamingParse helper. The "Parse (no schema)" baseline now measures
the streaming parser directly.

Net change: -1161 lines. 946 / 946 tests pass; benchmarks compile.
