+++ lib/stratiform/src/test/stratiform_test.scala
@@ -83,7 +83,7 @@ object Tests extends Suite(m"Stratiform Tests"):
       CorpusLoader.positive.each: testcase =>
         test(m"streaming parses ${testcase.stem}"):
           val cursor = Cursor[Data](testcase.source)
-          val doc = TelStreamParser.parse(cursor)
+          val doc = TelParser.parse(cursor)
           TelCheckTree.of(Tel.make(doc))
         . assert(_ == CheckFormat.parse(testcase.check).tree)
 
@@ -92,17 +92,17 @@ object Tests extends Suite(m"Stratiform Tests"):
         test(m"streaming matches TelParser on ${testcase.stem}"):
           val a = TelCheckTree.of(testcase.source.read[Tel])
           val b = TelCheckTree.of(Tel.make(
-            TelStreamParser.parse(Cursor[Data](testcase.source))))
+            TelParser.parse(Cursor[Data](testcase.source))))
           a == b
         . assert(identity)
 
     suite(m"Streaming parser — round-trip"):
       CorpusLoader.positive.each: testcase =>
         test(m"streaming round-trip ${testcase.stem}"):
-          val first = TelStreamParser.parse(Cursor[Data](testcase.source))
+          val first = TelParser.parse(Cursor[Data](testcase.source))
           val printed = Tel.show(Tel.make(first))
           val bytes: Data = summon[CharEncoder].encoded(printed)
-          val reparsed = TelStreamParser.parse(Cursor[Data](bytes))
+          val reparsed = TelParser.parse(Cursor[Data](bytes))
           TelCheckTree.of(Tel.make(reparsed)) == TelCheckTree.of(Tel.make(first))
         . assert(identity)
 
@@ -112,7 +112,7 @@ object Tests extends Suite(m"Stratiform Tests"):
         if codes.nonEmpty && codes.forall(_ < 200) then
           test(m"streaming raises an expected E1xx error on ${testcase.stem}"):
             codes.contains:
-              capture[TelError](TelStreamParser.parse(Cursor[Data](testcase.source)))
+              capture[TelError](TelParser.parse(Cursor[Data](testcase.source)))
               .reason.number
           . assert(_ == true)
 
@@ -131,11 +131,11 @@ object Tests extends Suite(m"Stratiform Tests"):
       CorpusLoader.positive.each: testcase =>
         test(m"all chunk sizes parse identically on ${testcase.stem}"):
           val baseline = TelCheckTree.of(Tel.make(
-            TelStreamParser.parse(Cursor[Data](testcase.source))))
+            TelParser.parse(Cursor[Data](testcase.source))))
           val sizes = List(1, 7, 64, 1024, testcase.source.length.max(1))
           sizes.forall: n =>
             val tree = TelCheckTree.of(Tel.make(
-              TelStreamParser.parse(chunkedCursor(testcase.source, n))))
+              TelParser.parse(chunkedCursor(testcase.source, n))))
             tree == baseline
         . assert(identity)
